move files to motech.repeaters
