+++ corehq/motech/repeaters/tests/test_migration_utils.py
@@ -1,7 +1,7 @@
 from django.test import TestCase
 
 from corehq.motech.repeaters.models import Repeater
-from corehq.motech.migration_utils import repair_repeaters_with_whitelist_bug
+from corehq.motech.repeaters.migration_utils import repair_repeaters_with_whitelist_bug
 
 
 class TestRepairRepeatersWithWhitelistBug(TestCase):
