+++ migrations.lock
@@ -668,7 +668,6 @@ motech
  0011_connectionsettings_is_deleted
  0012_alter_connectionsettings_auth_type
  0013_alter_connectionsettings_auth_type
- 0014_fix_whitelist_bug_repeaters
 notifications
  0001_squashed_0003_auto_20160504_2049 (3 squashed migrations)
  0002_auto_20160505_2058
@@ -779,6 +778,7 @@ repeaters
  0001_adjust_auth_field_format_squashed_0015_drop_connection_settings_fk
  0002_repeaters_db
  0003_id_fields
+ 0004_fix_whitelist_bug_repeaters
 reports
  0001_initial
  0002_auto_20171121_1803
